pyproject.toml
[tool.poetry]
name = "cooking-llm-agent"
version = "0.1.0"
description = "cooking llm agent"
authors = ["eshmelev", "tturushev", "rduan", "jsosa"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.13"
pydantic = "^2.12.5"
pydantic-settings = "^2.12.0"
tqdm = "^4.67.1"
async-lru = "^2.0.5"
langchain-core = "^1.0.4"
langchain-openai = "^1.0.3"
langgraph = "^1.0.3"
langgraph-checkpoint-postgres = "^2.0.0"
psycopg-pool = "^3.2.0"
python-dotenv = "^1.0.0"
arize-phoenix = ">=12.15.0,<13.0.0"
opentelemetry-instrumentation-httpx = "*"
openinference-instrumentation-langchain = "*"
fastapi = "^0.115.0"
uvicorn = "^0.32.0"
redis = "^5.0.0"


[tool.poetry.group.dev.dependencies]
black = "^25.11.0"
ruff = "^0.14.7"
mypy = "^1.19.0"
pytest = "^9.0.1"
ipykernel = "^7.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "UP"]
fix = true

[tool.mypy]
python_version = "3.11"
strict = false
ignore_missing_imports = true

[tool.poetry.scripts]
run_api = "src.api_handler.api_run:main"
agent_cli = "agent_cli:main"
backend_server = "backend.server:main"

[tool.pyright]
extraPaths = ["."]